+++ AllOfSoccer/Recruitment/Component/SearchPlaceView.swift
@@ -171,10 +171,7 @@ class SearchPlaceView: UIView {
     }
 
     private func setupInputPlaceTextField() {
-        let item1 = SearchTextFieldItem(title: "Blue", subtitle: "Color", image: UIImage(named: "icon_blue"))
-        let item2 = SearchTextFieldItem(title: "Red", subtitle: "Color", image: UIImage(named: "icon_red"))
-        let item3 = SearchTextFieldItem(title: "Yellow", subtitle: "Color", image: UIImage(named: "icon_yellow"))
-        self.inputPlaceTextField.filterItems([item1, item2, item3])
+        self.inputPlaceTextField.filterItems([])
     }
 
     @objc private func cancelButtonTouchUp(sender: UIButton) {
@@ -193,103 +190,14 @@ class SearchPlaceView: UIView {
         SearchPlaceAPI.request(searchText: searchText) { [weak self] result in
             switch result {
             case .success(let items):
-                print("wndgus: items \(items.map { $0.title })")
-
                 let filterItem = items.map { SearchTextFieldItem(title: $0.title) }
 
                 DispatchQueue.main.async {
                     self?.inputPlaceTextField.filterItems(filterItem)
-
                 }
             case .failure(let error):
                 print("Error: \(error.localizedDescription)")
             }
         }
     }
-}
-
-internal struct SearchPlaceAPI {
-
-    enum RequestError: Error {
-        case parsingError
-        case errorExist
-        case dataIsNil
-    }
-
-    static func request(searchText: String, completion: @escaping (Result<[Item], RequestError>) -> Void) {
-        let urlString = "https://openapi.naver.com/v1/search/local.json?query=\(searchText)&display=10".addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed) ?? ""
-        guard let url = URL(string: urlString) else {
-            return
-        }
-
-        var request = URLRequest(url: url)
-        request.addValue("IgvVcqCnm3_RhM7INYwu", forHTTPHeaderField: "X-Naver-Client-Id")
-        request.addValue("0FTVxW906A", forHTTPHeaderField: "X-Naver-Client-Secret")
-
-        URLSession.shared.dataTask(with: request) { data, response, error in
-            guard error == nil else {
-                completion(.failure(.errorExist))
-                return
-            }
-
-            guard let data = data else {
-                completion(.failure(.dataIsNil))
-                return
-            }
-
-            do {
-                let result = try JSONDecoder().decode(SearchPlaceAPI.ResponseValue.self, from: data)
-                completion(.success(result.items))
-            } catch {
-                completion(.failure(.parsingError))
-            }
-        }.resume()
-    }
-
-    struct ResponseValue: Codable {
-        let lastBuildDate: String
-        let total, start, display: Int
-        let items: [Item]
-
-        init(from decoder: Decoder) throws {
-            let container = try decoder.container(keyedBy: CodingKeys.self)
-
-            self.lastBuildDate = try container.decode(String.self, forKey: .lastBuildDate)
-            self.total = try container.decode(Int.self, forKey: .total)
-            self.start = try container.decode(Int.self, forKey: .start)
-            self.display = try container.decode(Int.self, forKey: .display)
-            self.items = try container.decode([Item].self, forKey: .items)
-        }
-    }
-
-    // MARK: - Item
-    struct Item: Codable {
-        let title: String
-        let link: String
-        let category, telephone, address: String
-        let roadAddress, mapx, mapy: String
-        let itemDescription: String?
-
-        enum CodingKeys: String, CodingKey {
-            case title, link, category
-            case itemDescription
-            case telephone, address, roadAddress, mapx, mapy
-        }
-
-        init(from decoder: Decoder) throws {
-            let container = try decoder.container(keyedBy: CodingKeys.self)
-
-            self.title = try container.decode(String.self, forKey: .title)
-            self.link = try container.decode(String.self, forKey: .link)
-            self.category = try container.decode(String.self, forKey: .category)
-            self.itemDescription = try? container.decode(String.self, forKey: .itemDescription)
-            self.address = try container.decode(String.self, forKey: .address)
-            self.roadAddress = try container.decode(String.self, forKey: .roadAddress)
-            self.mapx = try container.decode(String.self, forKey: .mapx)
-            self.mapy = try container.decode(String.self, forKey: .mapy)
-
-            self.telephone = try container.decode(String.self, forKey: .telephone)
-        }
-    }
-
 }
